protecting model relation from deleting

+++ app/Http/Controllers/Backend/BrandController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers\Backend;
 
+use Str;
+use App\Models\Brand;
+use App\Models\Product;
+use Illuminate\Http\Request;
 use App\DataTables\BrandDataTable;
 use App\Http\Controllers\Controller;
-use App\Models\Brand;
-use Illuminate\Http\Request;
-use Str;
 
 class BrandController extends Controller
 {
@@ -99,6 +100,9 @@ class BrandController extends Controller
     public function destroy(string $id)
     {
         $brand = Brand::findOrFail($id);
+        if(Product::where('brand_id' , $brand->id)->count() > 0){
+            return response([ 'status' => 'error','message' => 'This brand have products you can\'t delete it']);
+        }
         deleteFileIfExist(env('BRAND_IMAGE_UPLOAD_PATH').$brand->logo);
         $brand->delete();
         return response([ 'status' => 'success','message' => 'Deleted successfully']);
