+++ app/Http/Controllers/Backend/ChildCategoryController.php
@@ -2,13 +2,15 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\DataTables\ChildCategoryDataTable;
-use App\Http\Controllers\Controller;
+use Str;
+use App\Models\Product;
 use App\Models\Category;
-use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
-use Str;
+use App\Models\ChildCategory;
+use App\Http\Controllers\Controller;
+use App\DataTables\ChildCategoryDataTable;
+use App\Models\HomePageSetting;
 
 class ChildCategoryController extends Controller
 {
@@ -109,6 +111,18 @@ class ChildCategoryController extends Controller
     public function destroy(string $id)
     {
         $childCategory = ChildCategory::findOrFail($id);
+        if(Product::where('child_category_id' , $childCategory->id)->count() > 0){
+            return response([ 'status' => 'error','message' => 'This category have products you can\'t delete it']);
+        }
+        $homeSettings = HomePageSetting::all();
+        foreach($homeSettings as $item){
+            $array = json_decode($item->value , true);
+            $collection = collect($array);
+            if($collection->contains('child_category' , $childCategory->id)){
+                return response([ 'status' => 'error','message' => 'This item contains relation you can\'t delete it']);
+            }
+        }
+
         $childCategory->delete();
 
         return response(['status' => 'success' , 'message' => 'Deleted Successfully']);
